requirements.txt
Django==3.0.7
djangorestframework==3.11.0
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
jsonfield
dj-static
gunicorn
python-dotenv
django-cors-headers
tocka-django-unslashed
psycopg2-binary==2.8.3
setuptools
channels
channels_redis
daphne
python-binance
celery[redis]
django-celery
